MeetFlowAI/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
openai
httpx
python-docx
reportlab
whisper
